second

+++ Forum.Services/ArticlesManagement/ArticleService.cs
@@ -1,7 +1,8 @@
-﻿using System;
-using AutoMapper;
+﻿using AutoMapper;
 using Forum.Dal.Repository;
 using Forum.Models.ArticlesManagement;
+using Forum.WebApi.ErrorHandling;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -23,27 +24,40 @@ namespace Forum.Services.ArticlesManagement
             var article = _mapper.Map<ArticleRequest, Article>(articleRequest);
             if (article.Content.Length >= 2000)
             {
-                return null;
+                throw new ResponseException("articles bad content length", 400);
             }
 
             var articles = await _repository.GetAll();
             if (articles.Select(a => a.Title).Contains(articleRequest.Title))
             {
-                return null;
+                throw new ResponseException("article already exists with such title", 400);
             }
+            if (article == null)
+                throw new ResponseException("article wasn't created", 400);
+
 
             await _repository.Add(article);
             return article;
         }
 
-        public async Task DeleteEntityAsync(string articleId)
+        public async Task<bool> DeleteEntityAsync(Guid articleId)
         {
-            await _repository.Remove(articleId);
+            if ((await _repository.GetAll()).All(c => c.Id != articleId))
+            {
+                throw new ResponseException("there is no article with such id", 400);
+            }
+            return await _repository.Remove(articleId);
         }
 
-        public async Task<Article> GetEntityAsync(string articleId)
+        public async Task<Article> GetEntityAsync(Guid articleId)
         {
-            return await _repository.Find(articleId);
+            var article = await _repository.Find(articleId);
+            if (article == null)
+            {
+                throw new ResponseException("not found article", 404);
+            }
+
+            return article;
         }
 
         public async Task<IEnumerable<Article>> GetEntitiesAsync()
@@ -51,28 +65,36 @@ namespace Forum.Services.ArticlesManagement
             return await _repository.GetAll();
         }
 
-        public async Task<Article> UpdateEntityAsync(string articleId, ArticleRequest articleRequest)
+        public async Task<Article> UpdateEntityAsync(Guid articleId, ArticleRequest articleRequest)
         {
             var article = _mapper.Map<ArticleRequest, Article>(articleRequest);
-
-            await _repository.Update(article);
+            if (article.Content.Length >= 200)
+            {
+                throw new ResponseException("articles bad content length", 400);
+            }
+            
+            if (!await  _repository.Update(article))
+            {
+                throw new ResponseException("article did not update", 500);
+            }
             return article;
         }
 
         public async Task<IEnumerable<Article>> GetArticlesByOccurrenceAsync(string title, string userName, string category)
         {
             var articles = await _repository.GetAll();
-            var filterArticlesByTitle_and_User = articles?.Where(s => (title != null && s.Title.IndexOf(title) != -1) ||
+            var filterArticles = articles?.Where(s => (title != null && s.Title.IndexOf(title) != -1) ||
                                                       (userName != null && s.UserName.IndexOf(userName) != -1)).ToList();
 
-            category = category?.ToLower();
+            if (category != null)
+            {
+                category = category.ToLower();
+                var filterArticlesByCategories = articles?.Where(v => v.Category.ToString().ToLower().Contains(category)).ToList();
 
-            var filterArticlesByCategories= articles?.Where(v =>v.Category.ToString().ToLower().Contains(category)).ToList();
-            
-            var filterArticles = filterArticlesByTitle_and_User 
-                .Concat(filterArticlesByCategories).ToList();
-
-            return filterArticles.Count != 0 ? filterArticles : null;
+                filterArticles = filterArticles
+                   ?.Concat(filterArticlesByCategories).ToList();
+            }
+            return filterArticles?.Count != 0 ? filterArticles : null;
         }
     }
 }
